+++ lib/src/displayed_canvas.dart
@@ -192,7 +192,13 @@ class DisplayedCanvasState extends State<DisplayedCanvas>
       type: MaterialType.transparency,
       child: Stack(
         children: [
-          widget.child,
+          Listener(
+            behavior: HitTestBehavior.translucent,
+            onPointerDown: (event) => pointerMove(event.position),
+            onPointerMove: (event) => pointerMove(event.position),
+            onPointerUp: (event) => pointerUp(event.position),
+            child: widget.child,
+          ),
           IgnorePointer(
             child: AnimatedOpacity(
               duration: _theme.fadeDuration,
@@ -220,7 +226,7 @@ class DisplayedCanvasState extends State<DisplayedCanvas>
                         ),
                       ),
                     ),
-                  buildTooltipText(),
+                  _buildTooltipText(),
                   Flow(
                     delegate: PieDelegate(
                       bounceAnimation: _bounceAnimation,
@@ -260,7 +266,7 @@ class DisplayedCanvasState extends State<DisplayedCanvas>
     );
   }
 
-  Widget buildTooltipText() {
+  Widget _buildTooltipText() {
     double top;
     double dx = _pressedOffset.dx - _canvasOffset.dx;
     double dy = _pressedOffset.dy - _canvasOffset.dy;
@@ -356,20 +362,26 @@ class DisplayedCanvasState extends State<DisplayedCanvas>
     }
   }
 
+  bool isOutsideOfPointerArea(Offset offset) {
+    return (_pressedOffset - offset).distance > _theme.pointerSize / 2;
+  }
+
   void pointerDown({
     required Widget child,
     required RenderBox renderBox,
-    required Offset pressedOffset,
+    required Offset offset,
     required List<PieAction> actions,
     Function(bool menuVisible)? onMenuToggle,
     PieTheme? theme,
   }) {
-    if (!_pressed) {
+    if (_visible) {
+      pointerMove(offset);
+    } else if (!_pressed) {
       _onActiveMenuToggle = onMenuToggle;
       _theme = theme ?? widget.theme;
       _actions = actions;
       _pressed = true;
-      _pressedOffset = pressedOffset;
+      _pressedOffset = offset;
       _pointerDownTimer = Timer(_theme.delayDuration, () {
         _controller.forward(from: 0);
         setState(() {
@@ -383,13 +395,13 @@ class DisplayedCanvasState extends State<DisplayedCanvas>
     }
   }
 
-  void pointerUp() {
+  void pointerUp(Offset offset) {
     _pressed = false;
     if (_pointerDownTimer != null) {
       _pointerDownTimer!.cancel();
     }
 
-    if (_visible) {
+    if (_visible && isOutsideOfPointerArea(offset)) {
       if (_hoveredAction >= 0) {
         _actions[_hoveredAction].onSelect();
         Future.delayed(_theme.fadeDuration, () {
@@ -422,9 +434,8 @@ class DisplayedCanvasState extends State<DisplayedCanvas>
       if (_hoveredAction != -1) {
         setState(() => _hoveredAction = -1);
       }
-    } else if (_pressed &&
-        (offset - _pressedOffset).distance > _theme.pointerSize / 2) {
-      pointerUp();
+    } else if (_pressed && isOutsideOfPointerArea(offset)) {
+      pointerUp(offset);
     }
   }